requirements/prod.txt
git+https://github.com/potatolondon/djangae.git#egg=djangae
django>=1.6.10,<1.7
django-secure
django-csp
django-session-csrf
webassets
django_assets
git+https://github.com/potatolondon/django-enum.git
djangorestframework
